requirements.txt
python-telegram-bot~=21.10
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.3.0
Babel==2.9.1
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
class-registry==2.1.2
Django==5.1.4
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filters==1.3.2
fpdf==1.7.2
frozenlist==1.5.0
github==1.2.7
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
numpy==2.2.1
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2024.2
regex==2024.11.6
reportlab==4.2.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
yarl==1.18.3
